Do not create /var, /tmp, /usr and /home partitions
Having a single / partition is not nice for security reasons, since we
lose the ability to set different options for those file systems (like
nodev and nosuid), but it makes better use of the available disk space.

Since we don't know what role a generic FAI client will take on later,
how many users it will have, how much extra software will be installed,
and so on, I feel that this is the best default disk layout for now.